Refactoring of the project structure
