+++ src/service-orders/domain/interfaces/coordinate.interface.ts
